requirements.txt

openai-whisper
edge-tts
sounddevice
scipy
numpy
pygame
groq
python-dotenv
streamlit
plotly
pandas
streamlit-mic-recorder